+++ datatypes-cose/src/commonMain/kotlin/at/asitplus/crypto/datatypes/cose/CoseAlgorithm.kt
@@ -1,6 +1,6 @@
 package at.asitplus.crypto.datatypes.cose
 
-import at.asitplus.crypto.datatypes.JwsAlgorithm
+import at.asitplus.crypto.datatypes.CryptoAlgorithm
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.descriptors.PrimitiveKind
@@ -10,12 +10,12 @@ import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 
 @Serializable(with = CoseAlgorithmSerializer::class)
-enum class CoseAlgorithm(val value: Int) {
+enum class CoseAlgorithm(val value: Int, val isEC: Boolean = false) {
 
     // ECDSA with SHA-size
-    ES256(-7),
-    ES384(-35),
-    ES512(-36),
+    ES256(-7, true),
+    ES384(-35, true),
+    ES512(-36, true),
 
     // HMAC-size with SHA-size
     HS256(5),
@@ -32,46 +32,29 @@ enum class CoseAlgorithm(val value: Int) {
     RS384(-258),
     RS512(-259),
 
-    // Legacy support
+    // RSASSA-PKCS1-v1_5 using SHA-1
     RS1(-65535);
 
-    fun toJwsAlgorithm() = when(this) {
-        ES256 -> JwsAlgorithm.ES256
-        ES384 -> JwsAlgorithm.ES384
-        ES512 -> JwsAlgorithm.ES512
 
-        HS256 -> JwsAlgorithm.HS256
-        HS384 -> JwsAlgorithm.HS384
-        HS512 -> JwsAlgorithm.HS512
+    fun toCryptoAlgorithm() = when(this) {
+        ES256 -> CryptoAlgorithm.ES256
+        ES384 -> CryptoAlgorithm.ES384
+        ES512 -> CryptoAlgorithm.ES512
 
-        PS256 -> JwsAlgorithm.PS256
-        PS384 -> JwsAlgorithm.PS384
-        PS512 -> JwsAlgorithm.PS512
+        HS256 -> CryptoAlgorithm.HS256
+        HS384 -> CryptoAlgorithm.HS384
+        HS512 -> CryptoAlgorithm.HS512
 
-        RS256 -> JwsAlgorithm.RS256
-        RS384 -> JwsAlgorithm.RS384
-        RS512 -> JwsAlgorithm.RS512
-        RS1 -> JwsAlgorithm.NON_JWS_SHA1_WITH_RSA
+        PS256 -> CryptoAlgorithm.PS256
+        PS384 -> CryptoAlgorithm.PS384
+        PS512 -> CryptoAlgorithm.PS512
+
+        RS256 -> CryptoAlgorithm.RS256
+        RS384 -> CryptoAlgorithm.RS384
+        RS512 -> CryptoAlgorithm.RS512
+
+        RS1 -> CryptoAlgorithm.RS1
     }
-
-}
-fun JwsAlgorithm.toCoseAlgorithm() = when (this) {
-    JwsAlgorithm.ES256 -> CoseAlgorithm.ES256
-    JwsAlgorithm.ES384 -> CoseAlgorithm.ES384
-    JwsAlgorithm.ES512 -> CoseAlgorithm.ES512
-
-    JwsAlgorithm.HS256 -> CoseAlgorithm.HS256
-    JwsAlgorithm.HS384 -> CoseAlgorithm.HS384
-    JwsAlgorithm.HS512 -> CoseAlgorithm.HS512
-
-    JwsAlgorithm.PS256 -> CoseAlgorithm.PS256
-    JwsAlgorithm.PS384 -> CoseAlgorithm.PS384
-    JwsAlgorithm.PS512 -> CoseAlgorithm.PS512
-
-    JwsAlgorithm.RS256 -> CoseAlgorithm.RS256
-    JwsAlgorithm.RS384 -> CoseAlgorithm.RS384
-    JwsAlgorithm.RS512 -> CoseAlgorithm.RS512
-    JwsAlgorithm.NON_JWS_SHA1_WITH_RSA -> CoseAlgorithm.RS1
 }
 
 object CoseAlgorithmSerializer : KSerializer<CoseAlgorithm> {
@@ -88,4 +71,24 @@ object CoseAlgorithmSerializer : KSerializer<CoseAlgorithm> {
         return CoseAlgorithm.entries.first { it.value == decoded }
     }
 
+}
+
+fun CryptoAlgorithm.toCoseAlgorithm() = when(this) {
+    CryptoAlgorithm.ES256 -> CoseAlgorithm.ES256
+    CryptoAlgorithm.ES384 -> CoseAlgorithm.ES384
+    CryptoAlgorithm.ES512 -> CoseAlgorithm.ES512
+
+    CryptoAlgorithm.HS256 -> CoseAlgorithm.HS256
+    CryptoAlgorithm.HS384 -> CoseAlgorithm.HS384
+    CryptoAlgorithm.HS512 -> CoseAlgorithm.HS512
+
+    CryptoAlgorithm.PS256 -> CoseAlgorithm.PS256
+    CryptoAlgorithm.PS384 -> CoseAlgorithm.PS384
+    CryptoAlgorithm.PS512 -> CoseAlgorithm.PS512
+
+    CryptoAlgorithm.RS256 -> CoseAlgorithm.RS256
+    CryptoAlgorithm.RS384 -> CoseAlgorithm.RS384
+    CryptoAlgorithm.RS512 -> CoseAlgorithm.RS512
+
+    CryptoAlgorithm.RS1 -> CoseAlgorithm.RS1
 }