+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/CryptoSignature.kt
@@ -16,7 +16,7 @@ import kotlinx.serialization.encoding.Encoder
 
 /**
  * Data class which holds Asn1 Encoding of a signature of a specified algorithm
- * Allows simple ASN1 - DER - Raw transformation of signature values
+ * Allows simple ASN1 - Raw transformation of signature values
  * Does not check for anything!
  */
 
@@ -31,8 +31,6 @@ sealed class CryptoSignature(
      */
     abstract val rawByteArray: ByteArray
 
-
-    // abstract fun serialize(): String
     fun serialize(): String = signature.derEncoded.encodeToString(Base64UrlStrict)
 
     abstract fun encodeToTlvBitString(): Asn1Element
@@ -61,10 +59,10 @@ sealed class CryptoSignature(
             encoder.encodeString(value.serialize())
         }
     }
+
     /**
      * Input is expected to be x,y coordinates concatenated to bytearray
      */
-    //@Serializable(with = EC.CryptoSignatureSerializer::class)
     class EC(input: ByteArray) : CryptoSignature(
         asn1Sequence {
             append(
@@ -81,10 +79,6 @@ sealed class CryptoSignature(
             )
         }
     ) {
-//        override fun serialize(): String =
-//            //rawByteArray.encodeToString(Base64UrlStrict)
-//            signature.derEncoded.encodeToString(Base64UrlStrict)
-
         override val rawByteArray by lazy {
             val coordSizes = listOf(
                 32 - 1,
@@ -95,47 +89,21 @@ sealed class CryptoSignature(
                 coordSizes.filter { it <= ((signature as Asn1Sequence).children[0] as Asn1Primitive).content.size }
                     .minOrNull() ?: throw Exception("Illegal signature length")
             byteArrayOf(
-                *((signature as Asn1Sequence).children[0] as Asn1Primitive).decode(INTEGER) { it }.padWithZeros(coordSize),
+                *((signature as Asn1Sequence).children[0] as Asn1Primitive).decode(INTEGER) { it }
+                    .padWithZeros(coordSize),
                 *(signature.children[1] as Asn1Primitive).decode(INTEGER) { it }.padWithZeros(coordSize)
             )
         }
 
         override fun encodeToTlvBitString(): Asn1Element = encodeToDer().encodeToTlvBitString()
 
-//        object CryptoSignatureSerializer : KSerializer<EC> {
-//            override val descriptor: SerialDescriptor
-//                get() = PrimitiveSerialDescriptor("CryptoSignature", PrimitiveKind.STRING)
-//
-//            override fun deserialize(decoder: Decoder): EC =
-//                EC(decoder.decodeString().encodeToByteArray())
-//
-//            override fun serialize(encoder: Encoder, value: EC) {
-//                encoder.encodeString(value.serialize())
-//            }
-//        }
     }
 
-    //@Serializable(with = RSAorHMAC.CryptoSignatureSerializer::class)
     class RSAorHMAC(input: ByteArray) : CryptoSignature(
         Asn1Primitive(BIT_STRING, input)
     ) {
-//        override fun serialize(): String =
-//            rawByteArray.encodeToString(Base64UrlStrict)
-
         override val rawByteArray by lazy { (signature as Asn1Primitive).decode(BIT_STRING) { it } }
         override fun encodeToTlvBitString(): Asn1Element = this.encodeToTlv()
-
-//        object CryptoSignatureSerializer : KSerializer<RSAorHMAC> {
-//            override val descriptor: SerialDescriptor
-//                get() = PrimitiveSerialDescriptor("CryptoSignature", PrimitiveKind.STRING)
-//
-//            override fun deserialize(decoder: Decoder): RSAorHMAC =
-//                RSAorHMAC(decoder.decodeString().encodeToByteArray())
-//
-//            override fun serialize(encoder: Encoder, value: RSAorHMAC) {
-//                encoder.encodeString(value.serialize())
-//            }
-//        }
     }
 
     companion object : Asn1Decodable<Asn1Element, CryptoSignature> {
@@ -145,8 +113,12 @@ sealed class CryptoSignature(
                 when (src.tag) {
                     BIT_STRING -> RSAorHMAC((src as Asn1Primitive).decode(BIT_STRING) { it })
                     DERTags.DER_SEQUENCE -> {
-                        val first = ((src as Asn1Sequence).nextChild() as Asn1Primitive).decode(INTEGER) { it.dropWhile { it == 0.toByte() } }.toByteArray()
-                        val second = (src.nextChild() as Asn1Primitive).decode(INTEGER) { it.dropWhile { it == 0.toByte() } }.toByteArray()
+                        val first =
+                            ((src as Asn1Sequence).nextChild() as Asn1Primitive).decode(INTEGER) { it.dropWhile { it == 0.toByte() } }
+                                .toByteArray()
+                        val second =
+                            (src.nextChild() as Asn1Primitive).decode(INTEGER) { it.dropWhile { it == 0.toByte() } }
+                                .toByteArray()
                         if (src.hasMoreChildren()) throw IllegalArgumentException("Illegal Signature Format")
                         EC(first + second)
                     }