+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/pki/X509Certificate.kt
@@ -2,7 +2,7 @@ package at.asitplus.crypto.datatypes.pki
 
 import at.asitplus.crypto.datatypes.CryptoPublicKey
 import at.asitplus.crypto.datatypes.CryptoSignature
-import at.asitplus.crypto.datatypes.JwsAlgorithm
+import at.asitplus.crypto.datatypes.CryptoAlgorithm
 import at.asitplus.crypto.datatypes.asn1.*
 import at.asitplus.crypto.datatypes.asn1.DERTags.toImplicitTag
 import at.asitplus.crypto.datatypes.io.BitSet
@@ -17,7 +17,7 @@ import kotlinx.serialization.Serializable
 data class TbsCertificate(
     val version: Int = 2,
     @Serializable(with = ByteArrayBase64Serializer::class) val serialNumber: ByteArray,
-    val signatureAlgorithm: JwsAlgorithm,
+    val signatureAlgorithm: CryptoAlgorithm,
     val issuerName: List<DistinguishedName>,
     val validFrom: Asn1Time,
     val validUntil: Asn1Time,
@@ -122,7 +122,7 @@ data class TbsCertificate(
                 ((it as Asn1Tagged).verifyTag(0u).single() as Asn1Primitive).readInt()
             }
             val serialNumber = (src.nextChild() as Asn1Primitive).decode(BERTags.INTEGER) { it }
-            val sigAlg = JwsAlgorithm.decodeFromTlv(src.nextChild() as Asn1Sequence)
+            val sigAlg = CryptoAlgorithm.decodeFromTlv(src.nextChild() as Asn1Sequence)
             val issuerNames = (src.nextChild() as Asn1Sequence).children.map {
                 DistinguishedName.decodeFromTlv(it as Asn1Set)
             }
@@ -184,9 +184,10 @@ data class TbsCertificate(
 @Serializable
 data class X509Certificate(
     val tbsCertificate: TbsCertificate,
-    val signatureAlgorithm: JwsAlgorithm,
+    val signatureAlgorithm: CryptoAlgorithm,
     val signature: CryptoSignature
 ) : Asn1Encodable<Asn1Sequence> {
+
 
     @Throws(Asn1Exception::class)
     override fun encodeToTlv() = asn1Sequence {
@@ -222,9 +223,8 @@ data class X509Certificate(
         @Throws(Asn1Exception::class)
         override fun decodeFromTlv(src: Asn1Sequence): X509Certificate = runRethrowing {
             val tbs = TbsCertificate.decodeFromTlv(src.nextChild() as Asn1Sequence)
-            val sigAlg = JwsAlgorithm.decodeFromTlv(src.nextChild() as Asn1Sequence)
-            val test = src.nextChild()
-            val signature = CryptoSignature.decodeFromTlv(test)
+            val sigAlg = CryptoAlgorithm.decodeFromTlv(src.nextChild() as Asn1Sequence)
+            val signature = CryptoSignature.decodeFromTlv(src.nextChild())
             if (src.hasMoreChildren()) throw Asn1StructuralException("Superfluous structure in Certificate Structure")
             return X509Certificate(tbs, sigAlg, signature)
         }