+++ CHANGELOG.md
@@ -39,6 +39,6 @@
 * Change `JwsHeader.publicKey` from JsonWebKey to CryptoPublicKey
 * Remove `SignatureValueLength` parameters from JWS- & Cose Algorithm Enum class
 * Remove deprecated functions
-* New `JwsAlgorithm` to `CoseAlgorithm` conversion
+* New `CryptoAlgorithm` class 
 * New `CryptoSignature` class for easy Asn1 - RawByteArray conversion
 * Rename function `.toJcaPublicKey` from `.toPublicKey` to reflect connection to JVM